Add kill-window script
Sometimes, i3 is no longer able to close an opened window. The
corresponding window is usually freezed and no longer usable. i3's
default methods of killing it do not work in this state.

The kill-window script allows to select an open window within rofi,
which is when killed via xkill.