Update i3-flameshot
With newer versions of flameshot, the window create event of the
screenshot save menu was only catched about 50% of the time. I guess the
time between flameshot gui being closed and the save menu to popup was
to short to start the event listener, but not sure. This commit solves
the problem by using Popen instead of call, to start flameshot in the
background. The wait time for the screenshot save window was increased
to 2min.